% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_u_colors}
\alias{get_u_colors}
\title{returns list of CI/CD color values}
\usage{
get_u_colors()
}
\value{
list of hex colors
}
\description{
returns list of CI/CD color values
}
\examples{
u_colors()

}
